LayoutTests/crypto/subtle/hmac-export-key-expected.txt

Test exporting an AES key.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a JWK key...
Exporting the key as raw data...
PASS byteArrayToHexString(new Uint8Array(exportedData)) is '[6a 18 e4 9f ef f7 f3 b7 e0 9e c8 9b 7f 6d ea b2 f6 a1 8e 49 fe ff 7f 3b 7e 09 ec 89 b7 f6 de ab]'
PASS successfullyParsed is true

TEST COMPLETE
